Refactor to use ddf-test as test artifacts, test classes are implemented at ddf-on-x projects
