+++ ddf-test/src/main/resources/log4j.properties
