+++ ddf-test/src/main/scala/io/ddf/test/it/AggregationSpec.scala
@@ -20,6 +20,7 @@ package io.ddf.test.it
 
 import io.ddf.types.AggregateTypes.AggregateFunction
 import org.scalatest.Matchers
+
 import scala.collection.JavaConversions._
 
 trait AggregationSpec extends BaseSpec with Matchers {
