--- ddf-test/src/main/scala/io/ddf/test/it/Main.scala
@@ -1,8 +1,0 @@
-package scala.io.ddf.test.it
-
-/**
-  * Created by hainguyen on 1/12/16.
-  */
-class Main {
-
-}
