+++ ddf-test/src/main/scala/io/ddf/test/it/SchemaSpec.scala
@@ -18,8 +18,8 @@
  */
 package io.ddf.test.it
 
-import org.scalatest.Matchers
 import io.ddf.content.Schema.{Column, ColumnClass, ColumnType}
+import org.scalatest.Matchers
 
 import scala.collection.JavaConverters._
