+++ spark/pom.xml
@@ -216,6 +216,12 @@
         </dependency>
         <dependency>
             <groupId>io.ddf</groupId>
+            <artifactId>ddf_test_2.10</artifactId>
+            <version>1.4.9-SNAPSHOT</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.ddf</groupId>
             <artifactId>ddf_core_2.10</artifactId>
             <version>1.4.9-SNAPSHOT</version>
         </dependency>
